+++ docs/releasenotes.rst
@@ -3,7 +3,8 @@ Release notes
 
 Version 0.1.8
 -------------
-- added possibility to export to `zarr <https://github.com/zarr-developers/zarr>`__
+- create and open arbitrary (non-protected) files in darr array directory
+- export darr to `zarr <https://github.com/zarr-developers/zarr>`__
 - asarray works on zarr arrays more efficiently
 - added support for Maple
 - removed set_accessmode method, now set accessmode attribute directly